greentea/tgvisd: Refactor Scraper.cpp
